+++ lib/db/src/schema/index.ts
@@ -1,3 +1,4 @@
 export * from "./documents";
 export * from "./chunks";
 export * from "./chat_messages";
+export * from "./users";